simplify store interface (#190)
* simplify store interface

* changed interface somewhat

* removed outbox stuff

* restored generic_test_accepts_only_once

* actually call the callback

+++ python/gink/impl/log_backed_store.py
@@ -1,19 +1,23 @@
 """ implementation of the LogBackedStore class """
-from typing import Tuple
-from fcntl import flock, LOCK_EX, LOCK_NB
+from typing import Optional, Callable, Union
+from fcntl import flock, LOCK_EX, LOCK_NB, LOCK_UN
 from .builders import LogFile
 from .memory_store import MemoryStore
 from .bundle_info import BundleInfo
+from .bundle_wrapper import BundleWrapper
+
 
 
 class LogBackedStore(MemoryStore):
     """A Store backed by a simple append-only file."""
 
-    def __init__(self, filepath, reset=False):
+    def __init__(self, filepath, *, exclusive=False, reset=False):
         MemoryStore.__init__(self)
         self._filepath = filepath
         self._handle = open(self._filepath, "ab+")
-        flock(self._handle, LOCK_EX | LOCK_NB)
+        self._exclusive = exclusive
+        if self._exclusive:
+            flock(self._handle, LOCK_EX | LOCK_NB)  # this will throw if another process has a lock
         if reset:
             self._handle.truncate()
         self._handle.seek(0)
@@ -22,17 +26,28 @@ class LogBackedStore(MemoryStore):
         for bundle_bytes in self._log_file_builder.commits:  # type: ignore # pylint: disable=maybe-no-member
             MemoryStore.apply_bundle(self, bundle_bytes=bundle_bytes)
 
-    def apply_bundle(self, bundle_bytes: bytes, push_into_outbox: bool = False) -> Tuple[BundleInfo, bool]:
+    def refresh(self):
+        pass
+
+    def apply_bundle(self, bundle: Union[BundleWrapper, bytes], callback: Optional[Callable]=None) -> bool:
         if self._handle.closed:
             raise AssertionError("attempt to write to closed LogBackStore")
-        bundle_info, added = MemoryStore.apply_bundle(self, bundle_bytes, push_into_outbox)
+        if isinstance(bundle, bytes):
+            bundle = BundleWrapper(bundle)
+        if not self._exclusive:
+            flock(self._handle, LOCK_EX)  # this will block (wait) if another process has a lock
+        added = MemoryStore.apply_bundle(self, bundle)
         if added:
             self._log_file_builder.Clear()  # type: ignore
             self._log_file_builder.commits.push(bundle_bytes)  # type: ignore
             data: bytes = self._log_file_builder.SerializeToString()  # type: ignore
             self._handle.write(data)
             self._handle.flush()
-        return bundle_info, added
+            if callback is not None:
+                callback(bundle)
+        if not self._exclusive:
+            flock(self._handle, LOCK_UN)
+        return added
 
     def get_claimed_chains(self):
         raise NotImplementedError()
